--- app/src/main/java/com/example/projecteyebrow/common/Constant.kt
@@ -1,6 +1,0 @@
-package com.example.projecteyebrow.common
-
-object Constant {
-    const val MAX_RETRY_ATTEMPT: Int = 3
-    const val DELAY_TIME_MILLIS: Long = 1000L
-}
