+++ app/src/main/java/com/example/projecteyebrow/module/ToastModule.kt
@@ -1,8 +1,7 @@
-package com.example.projecteyebrow.di.module
+package com.example.projecteyebrow.module
 
 import android.content.Context
-import com.bumptech.glide.Glide
-import com.bumptech.glide.RequestManager
+import android.widget.Toast
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -10,11 +9,11 @@ import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import javax.inject.Singleton
 
-/* Glide ImageLoader Module */
 @Module
 @InstallIn(SingletonComponent::class)
-object GlideModule {
+object ToastModule {
     @Provides
     @Singleton
-    fun providesGlideModule(@ApplicationContext context: Context):RequestManager = Glide.with(context)
+    fun providesToast(@ApplicationContext context: Context): Toast =
+        Toast.makeText(context, "", Toast.LENGTH_SHORT)
 }
