+++ app/src/main/java/com/example/projecteyebrow/view/CommunityFragment.kt
@@ -11,7 +11,7 @@ import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
 import com.example.projecteyebrow.R
 import com.example.projecteyebrow.databinding.FragmentCommunityBinding
-import com.example.projecteyebrow.di.dispatcherQualifier.MainDispatcher
+import com.example.projecteyebrow.Qualifier.MainDispatcher
 import com.example.projecteyebrow.view.community.CommunityContentList
 import com.example.projecteyebrow.viewModel.CommunityViewModel
 import dagger.hilt.android.AndroidEntryPoint
@@ -21,7 +21,8 @@ import javax.inject.Inject
 
 @AndroidEntryPoint
 class CommunityFragment : Fragment() {
-    @Inject @MainDispatcher lateinit var mainDispatcher: CoroutineDispatcher
+    @Inject @MainDispatcher
+    lateinit var mainDispatcher: CoroutineDispatcher
     @Inject lateinit var toastMessage: Toast
 
     private var _binding: FragmentCommunityBinding? = null
