+++ app/src/main/java/com/example/projecteyebrow/viewModel/CommunityViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.viewModelScope
 import com.example.domain.entity.CommunityItem
 import com.example.domain.usecase.auth.GetCurrentUserSessionUseCase
 import com.example.domain.usecase.fireDB.ReadAllCommunityContentUseCase
-import com.example.projecteyebrow.di.dispatcherQualifier.IoDispatcher
+import com.example.projecteyebrow.Qualifier.IoDispatcher
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.cancel
