+++ app/src/main/java/com/example/projecteyebrow/viewModel/ProfileViewModel.kt
@@ -8,7 +8,7 @@ import com.example.domain.entity.ProfileItem
 import com.example.domain.usecase.auth.LogOutUserAccountUseCase
 import com.example.domain.usecase.fireDB.LoadUserProfileUseCase
 import com.example.domain.usecase.fireDB.StopEventListenUseCase
-import com.example.projecteyebrow.di.dispatcherQualifier.IoDispatcher
+import com.example.projecteyebrow.Qualifier.IoDispatcher
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Job
