+++ app/src/main/java/com/example/projecteyebrow/viewModel/WriteContentViewModel.kt
@@ -7,7 +7,7 @@ import androidx.lifecycle.viewModelScope
 import com.example.domain.entity.TemporaryCommunityItem
 import com.example.domain.usecase.fireDB.UploadCommunityContentUseCase
 import com.example.domain.usecase.roomDB.SaveTempCommunityItemUseCase
-import com.example.projecteyebrow.di.dispatcherQualifier.IoDispatcher
+import com.example.projecteyebrow.Qualifier.IoDispatcher
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Job
